+++ models/forms/CalendarEntryForm.php
@@ -9,6 +9,8 @@
 
 namespace humhub\modules\calendar\models\forms;
 
+use humhub\modules\space\models\Space;
+use humhub\modules\topic\models\Topic;
 use Yii;
 use yii\base\Model;
 use DateInterval;
@@ -67,9 +69,24 @@ class CalendarEntryForm extends Model
     public $type_id;
 
     /**
+     * @var
+     */
+    public $topics = [];
+
+    /*
+     * @var array
+     */
+    public $markdownFiles = [];
+
+    /**
      * @var bool
      */
     public $sendUpdateNotification = 0;
+
+    /**
+     * @var integer if set to true all space participants will be added to the event
+     */
+    public $forceJoin = 0;
 
     /**
      * @var CalendarEntry
@@ -93,6 +110,8 @@ class CalendarEntryForm extends Model
             if(!empty($type)) {
                 $this->type_id = $type->id;
             }
+
+            $this->topics = $this->entry->content->getTags(Topic::class);
         }
     }
 
@@ -103,10 +122,11 @@ class CalendarEntryForm extends Model
     {
         return [
             [['timeZone'], 'in', 'range' => DateTimeZone::listIdentifiers()],
-            [['is_public', 'type_id', 'sendUpdateNotification'], 'integer'],
+            [['topics'], 'safe'],
+            [['is_public', 'type_id', 'sendUpdateNotification', 'forceJoin'], 'integer'],
             [['start_time', 'end_time'], 'date', 'type' => 'time', 'format' => $this->getTimeFormat()],
-            [['start_date'], DbDateValidator::className(), 'format' => Yii::$app->params['formatter']['defaultDateFormat'], 'timeAttribute' => 'start_time', 'timeZone' => $this->timeZone],
-            [['end_date'], DbDateValidator::className(), 'format' => Yii::$app->params['formatter']['defaultDateFormat'], 'timeAttribute' => 'end_time', 'timeZone' => $this->timeZone],
+            [['start_date'], DbDateValidator::class, 'format' => Yii::$app->params['formatter']['defaultDateFormat'], 'timeAttribute' => 'start_time', 'timeZone' => $this->timeZone],
+            [['end_date'], DbDateValidator::class, 'format' => Yii::$app->params['formatter']['defaultDateFormat'], 'timeAttribute' => 'end_time', 'timeZone' => $this->timeZone],
             [['end_date'], 'validateEndTime'],
             [['type_id'], 'validateType'],
         ];
@@ -174,6 +194,9 @@ class CalendarEntryForm extends Model
             'timeZone' => Yii::t('CalendarModule.base', 'Time Zone'),
             'is_public' => Yii::t('CalendarModule.base', 'Public'),
             'sendUpdateNotification' => Yii::t('CalendarModule.base', 'Send update notification'),
+            'forceJoin' => ($this->entry->isNewRecord)
+                ? Yii::t('CalendarModule.base', 'Add all space members to this event')
+                : Yii::t('CalendarModule.base', 'Add remaining space members to this event'),
         ];
     }
 
@@ -203,7 +226,6 @@ class CalendarEntryForm extends Model
             $this->entry->time_zone = $this->timeZone;
         }
 
-        //$this->translateDateTimes($this->entry->start_datetime, $this->entry->end_datetime, Yii::$app->timeZone, $this->timeZone);
 
         $this->entry->content->visibility = $this->is_public;
 
@@ -219,6 +241,10 @@ class CalendarEntryForm extends Model
         return true;
     }
 
+    /**
+     * @return bool
+     * @throws \Throwable
+     */
     public function save()
     {
         if(!$this->validate()) {
@@ -232,20 +258,29 @@ class CalendarEntryForm extends Model
         // The form expects user time zone, so we translate back from app to user timezone
         $this->translateDateTimes($this->entry->start_datetime, $this->entry->end_datetime, Yii::$app->timeZone, $this->timeZone);
 
-        if($this->entry->save()) {
-            $this->entry->fileManager->attach($this->entry->files);
-            if(!empty($this->type_id)) {
-                $this->entry->setType($this->type_id);
+        return CalendarEntry::getDb()->transaction(function($db) {
+            if($this->entry->save()) {
+                $this->entry->fileManager->attach($this->entry->files);
+
+                if(!empty($this->type_id)) {
+                    $this->entry->setType($this->type_id);
+                }
+
+                if($this->sendUpdateNotification && !$this->entry->isNewRecord) {
+                    $this->entry->sendUpdateNotification();
+                }
+
+                if($this->forceJoin) {
+                    $this->entry->addAllUsers();
+                }
+
+                Topic::attach($this->entry->content, $this->topics);
+
+                return true;
             }
 
-            if($this->sendUpdateNotification && !$this->entry->isNewRecord) {
-                $this->entry->sendUpdateNotification();
-            }
-
-            return true;
-        }
-
-        return false;
+            return false;
+        });
     }
 
     public static function getParticipationModeItems()