+++ notifications/EventUpdated.php
@@ -65,7 +65,7 @@ class EventUpdated extends BaseNotification
     public function getMailSubject()
     {
         return Yii::t('CalendarModule.notifications_views_CanceledEvent', '{displayName} just updated event {contentTitle}.', [
-            'displayName' => Html::tag('strong', Html::encode($this->originator->displayName)),
+            'displayName' => Html::encode($this->originator->displayName),
             'contentTitle' => $this->getContentInfo($this->source, false)
         ]);
     }