+++ notifications/ForceAdd.php
@@ -1,0 +1,65 @@
+<?php
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
+ *
+ */
+
+namespace  humhub\modules\calendar\notifications;
+
+use humhub\libs\Html;
+use humhub\modules\content\notifications\ContentCreatedNotificationCategory;
+use humhub\modules\notification\components\BaseNotification;
+use humhub\modules\space\models\Space;
+use Yii;
+
+/**
+ * Created by PhpStorm.
+ * User: buddha
+ * Date: 21.07.2017
+ * Time: 23:12
+ */
+class ForceAdd extends BaseNotification
+{
+    /**
+     * @inheritdoc
+     */
+    public $moduleId = 'calendar';
+
+    /**
+     * @inheritdoc
+     */
+    public $viewName = 'participationInfoNotification';
+
+    /**
+     * @inheritdoc
+     */
+    public function category()
+    {
+        return new CalendarNotificationCategory();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function html()
+    {
+        return Yii::t('CalendarModule.base', '{displayName} just added you to event {contentTitle} in space {spaceName}.', [
+            'displayName' => Html::tag('strong', Html::encode($this->originator->displayName)),
+            'contentTitle' => $this->getContentInfo($this->source, false),
+            'spaceName' =>  Html::encode($this->source->content->container->displayName)
+        ]);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getMailSubject()
+    {
+        return Yii::t('CalendarModule.notifications_views_CanceledEvent', '{displayName} just added you to event {contentTitle}.', [
+            'displayName' =>  Html::encode($this->originator->displayName),
+            'contentTitle' => $this->getContentInfo($this->source, false)
+        ]);
+    }
+}